feat: make raid list command to remove emeralds or aspects from provided user

+++ ogybot.Domain/Entities/UserTypes/RaidListUser.cs
@@ -6,7 +6,7 @@ public class RaidListUser : User
 {
 
     public double Aspects { get; set; }
-    public int EmeraldsOwed { get; set; }
+    public double EmeraldsOwed { get; set; }
     public int Raids { get; set; }
 
     public RaidListUser(string username) : base(username)
